Fix compilation warnings when using clang

+++ tests/check/element/passthrough.c
@@ -79,7 +79,7 @@ fakesink_hand_off_check_size (GstElement * fakesink, GstBuffer * buf,
     // TODO: Count for size
     GST_INFO ("Bitrate is: %ld bits", bitrate);
 
-    fail_if (abs (bitrate - BITRATE) > BITRATE * 0.15);
+    fail_if (abs ((int) ((gssize) bitrate - BITRATE)) > BITRATE * 0.15);
     g_object_set (G_OBJECT (fakesink), "signal-handoffs", FALSE, NULL);
     g_idle_add (quit_main_loop_idle, loop);
   }
